// src/compute/types.generated.ts
export interface CalculatorInputsById {
// ...
  "nihss": {
    "assessment_timing": "baseline" | "post_treatment" | "follow_up";
    "modifier_context": "none" | "intubation" | "amputation_or_joint_fusion" | "language_barrier" | "pre_existing_deficit" | "other" | "multiple_modifiers";
    "level_of_consciousness": "alert" | "minor_stimulation" | "repeated_stimulation" | "movements_to_pain" | "unresponsive";
    "month_and_age": "both_right" | "one_right" | "none_right" | "impaired_communication" | "aphasic";
    "commands": "both_tasks" | "one_task" | "no_tasks";
    "horizontal_gaze": "normal" | "partial_overcome" | "partial_reflex" | "forced_palsy";
    "visual_fields": "no_loss" | "partial_hemianopia" | "complete_hemianopia" | "bilaterally_blind" | "bilateral_hemianopia";
    "facial_palsy": "normal" | "minor_paralysis" | "partial_paralysis" | "unilateral_complete" | "bilateral_complete";
    "left_arm_motor": "no_drift" | "drift_no_bed" | "drift_hits_bed" | "some_effort" | "no_effort" | "no_movement" | "amputation";
    "right_arm_motor": "no_drift" | "drift_no_bed" | "drift_hits_bed" | "some_effort" | "no_effort" | "no_movement" | "amputation";
    "left_leg_motor": "no_drift" | "drift_no_bed" | "drift_hits_bed" | "some_effort" | "no_effort" | "no_movement" | "amputation";
    "right_leg_motor": "no_drift" | "drift_no_bed" | "drift_hits_bed" | "some_effort" | "no_effort" | "no_movement" | "amputation";
    "limb_ataxia": "no_ataxia" | "one_limb" | "two_limbs" | "does_not_understand" | "paralyzed" | "amputation";
    "sensation": "normal" | "mild_moderate_sharp" | "mild_moderate_sense" | "complete_loss" | "quadriplegic" | "unresponsive";
    "language": "normal" | "mild_moderate" | "severe" | "mute" | "unresponsive";
    "dysarthria": "normal" | "mild_moderate" | "severe" | "mute" | "intubated";
    "extinction": "no_abnormality" | "inattention" | "extinction" | "profound_hemi_inattention" | "multiple_modalities";
  };
  "oxygenation_index": {
    "mean_airway_pressure": number;
    "fio2": number;
    "pao2"?: number;
    "spo2"?: number;
  };
  "pews": {
    "behavior": "playing_appropriate" | "sleeping" | "irritable" | "lethargic";
    "cardiovascular": "pink" | "pale" | "gray" | "gray_mottled";
    "respiratory": "normal" | "mild_distress" | "moderate_distress" | "severe_distress";
    "nebulizers": boolean;
    "vomiting": boolean;
  };
  "qsofa": {
    "respiratory_rate": number;
    "systolic_bp": number;
    "altered_mental_status": boolean;
  };
  "r_factor": {
    "alt": number;
    "alp": number;
    "alt_uln": number;
    "alp_uln": number;
  };
  "ranson": {
    "etiology_variant": "non_gallstone" | "gallstone";
    "assessment_hours": number;
    "admission_observed_at_hours": number;
    "followup_observed_at_hours"?: number;
    "age_years": number;
    "wbc_admission": number;
    "glucose_admission": number;
    "ast_admission": number;
    "ldh_admission": number;
    "hematocrit_admission": number;
    "bun_admission": number;
    "hematocrit_followup"?: number;
    "bun_followup"?: number;
    "calcium_followup"?: number;
    "pao2_followup"?: number;
    "base_deficit_followup"?: number;
    "fluid_sequestration_followup"?: number;
  };
// ...
  "wells_dvt": {
    "active_cancer": boolean;
    "paralysis_or_cast": boolean;
    "bedridden": boolean;
    "localized_tenderness": boolean;
    "entire_leg_swollen": boolean;
    "calf_swelling": boolean;
    "pitting_edema": boolean;
    "collateral_veins": boolean;
    "alternative_diagnosis": boolean;
    "previously_diagnosed": boolean;
  };
}